+++ tsconfig.app.tsbuildinfo
@@ -1,1 +1,1 @@
-{"root":["./src/app.tsx","./src/main.tsx","./src/type.d.ts","./src/vite-env.d.ts","./src/components/mswwrapper.tsx","./src/components/chats/chat-item.tsx","./src/components/chats/chat-list.tsx","./src/components/chats/chat-section.tsx","./src/components/messages/input-field.tsx","./src/components/messages/message-item.tsx","./src/components/messages/message-list.tsx","./src/components/messages/message-section.tsx","./src/components/pages/chat.tsx","./src/hooks/api.ts","./src/hooks/chats/useaddchat.ts","./src/hooks/chats/usechatlist.ts","./src/hooks/chats/usechatmodellist.ts","./src/hooks/messages/useaddmessage.ts","./src/hooks/messages/usemessagelist.ts","./src/mock/browser.ts","./src/mock/data.ts","./src/mock/handlers.ts","./src/mock/type.ts","./src/shared/buttons/component.tsx","./src/shared/buttons/type.ts","./src/shared/dropdown/body.tsx","./src/shared/dropdown/component.tsx","./src/shared/dropdown/head.tsx","./src/shared/dropdown/hook.ts","./src/shared/dropdown/item.tsx","./src/shared/dropdown/type.ts","./src/shared/inputs/textarea/component.tsx","./src/shared/inputs/textarea/type.ts","./src/shared/loading/loading.tsx","./src/shared/profile/profile.tsx","./src/stores/chat/chatstore.tsx","./src/stores/chat/type.ts","./src/stores/query/global.tsx","./src/utils/common.ts"],"version":"5.6.3"}
+{"root":["./src/app.tsx","./src/main.tsx","./src/type.d.ts","./src/vite-env.d.ts","./src/components/mswwrapper.tsx","./src/components/chats/chat-item.tsx","./src/components/chats/chat-list.tsx","./src/components/chats/chat-section.tsx","./src/components/messages/input-field.tsx","./src/components/messages/message-header.tsx","./src/components/messages/message-item.tsx","./src/components/messages/message-list.tsx","./src/components/messages/message-section.tsx","./src/components/pages/chat.tsx","./src/mock/browser.ts","./src/mock/data.ts","./src/mock/handlers.ts","./src/mock/type.ts","./src/services/api.ts","./src/services/chats/useaddchat.ts","./src/services/chats/usechatlist.ts","./src/services/chats/usechatmodellist.ts","./src/services/messages/useaddmessage.ts","./src/services/messages/usemessagelist.ts","./src/shared/buttons/component.tsx","./src/shared/buttons/type.ts","./src/shared/dropdown/body.tsx","./src/shared/dropdown/component.tsx","./src/shared/dropdown/head.tsx","./src/shared/dropdown/hook.ts","./src/shared/dropdown/item.tsx","./src/shared/dropdown/type.ts","./src/shared/inputs/textarea/component.tsx","./src/shared/inputs/textarea/type.ts","./src/shared/loading/loading.tsx","./src/shared/profile/profile.tsx","./src/stores/chat/chatstore.tsx","./src/stores/chat/type.ts","./src/stores/query/global.tsx","./src/utils/common.ts"],"version":"5.6.3"}
